+++ src/pages/calendar/Calendar.jsx
@@ -11,9 +11,9 @@ import { callDeleteCalendarAPI, callInsertCalendarAPI, callSelectCalendarAPI, ca
 import { useSelector, useDispatch } from 'react-redux';
 import '../../css/common.css'
 import CalendarUpdateModal from './CalendarUpdateModal';
-import { Popover } from 'bootstrap';
-import { convertToUtc } from '../../utils/CalendarUtill';
+import { calendarPopover, convertToUtc, updateEvents } from '../../utils/CalendarUtill';
 import { Link } from 'react-router-dom';
+import { decodeJwt } from '../../utils/tokenUtils';
 
 
 function Calendar() {
@@ -22,6 +22,7 @@ function Calendar() {
     const [events, setEvents] = useState([]);
     const [selectedEvent, setSelectedEvent] = useState();
     const dispatch = useDispatch();
+    const token = decodeJwt(window.localStorage.getItem("accessToken"));
 
 
     const handleOpenModal = () => {
@@ -33,7 +34,7 @@ function Calendar() {
         setSelectedEvent();
     };
 
-    const handleEventClick = (info) => {
+    const handleEventClick = info => {
         setSelectedEvent(info.event);
         handleOpenModal();
     };
@@ -46,10 +47,9 @@ function Calendar() {
             color,
             department: "개발팀", // TODO: 나중에 부서 선택해서 추가하는 기능넣어야 함
             detail,
-            registrantId: 200401023 // TODO: 나중에 관리자 사번 뽑아서 넣는 걸로 바꿔야 함
+            registrantId: token.memberId
         };
         dispatch(callInsertCalendarAPI(requestData));
-
     };
 
     const handleUpdateChanges = ({ id, title, start, end, color, detail }) => {
@@ -62,55 +62,14 @@ function Calendar() {
             detail
         };
         dispatch(callUpdateCalendarAPI(requestData));
-
     };
 
-    const handleDeleteChanges = (id) => {
-        dispatch(callDeleteCalendarAPI(id));
-    }
+    const handleDeleteChanges = id => dispatch(callDeleteCalendarAPI(id));
 
-
-    useEffect(() => {
-        const department = "개발팀"; // TODO: 나중에 부서 선택해서 추가하는 기능넣어야 함
-        dispatch(callSelectCalendarAPI(department));
-
-    }, []);
-
-    useEffect(() => {
-        const department = "개발팀"; // TODO: 나중에 부서 선택해서 추가하는 기능넣어야 함
-        dispatch(callSelectCalendarAPI(department));
-
-    }, [insertMessage, updateMessage, deleteMessage]);
-
-    useEffect(() => {
-        if (calendarList) {
-            const newEvents = calendarList.map(calendar => {
-                // Javascript date 타입의 month 가 0~11 db에 저장된 월에서 1을 빼줘야 정상적으로 표현됨
-                calendar.calendarStart[1] = calendar.calendarStart[1] - 1;
-                calendar.calendarEnd[1] = calendar.calendarEnd[1] - 1;
-
-                // 배경이 노랑이면 흰색글씨가 안보여서 노랑일 때 글씨색 변경
-                let textColor = 'white';
-                if (calendar.color === 'yellow') {
-                    textColor = 'black';
-                }
-
-                return {
-                    id: calendar.calendarNo,
-                    title: calendar.calendarName,
-                    start: calendar.calendarStart,
-                    end: calendar.calendarEnd,
-                    color: calendar.color,
-                    textColor: textColor,
-                    extendedProps: {
-                        color: calendar.color,
-                        detail: calendar.detail
-                    }
-                }
-            });
-            setEvents(newEvents);
-        }
-    }, [calendarList]);
+    // TODO: 나중에 부서 선택해서 추가하는 기능넣어야 함
+    useEffect(() => {dispatch(callSelectCalendarAPI("개발팀"))} , [insertMessage, updateMessage, deleteMessage]);
+    
+    useEffect(() => updateEvents(calendarList, setEvents), [calendarList]);
 
 
     return <>
@@ -140,27 +99,7 @@ function Calendar() {
                 }}
 
                 events={events}
-                eventDidMount={(info) => {
-                    let content = `<p>시작일시: ${info.event.start.toLocaleString()}</p>`;
-                    if (info.event.end) {
-                        content += `<p>종료일시: ${info.event.end.toLocaleString()}</p>`;
-                    } else {
-                        content += '<p>종료일: 당일'
-                    }
-                    if (info.event.extendedProps.detail) {
-                        content += `<p>일정상세: ${info.event.extendedProps.detail}</p>`
-                    } else {
-                        content += `<p>일정상세: 없음</p>`
-                    }
-                    return new Popover(info.el, {
-                        title: info.event.title,
-                        placement: "auto",
-                        trigger: "hover",
-                        customClass: "popoverStyle",
-                        content: content,
-                        html: true,
-                    });
-                }}
+                eventDidMount={info => calendarPopover(info)}
                 eventClick={handleEventClick}
                 locale={koLocale}
             />
